Use embedded icons for money loot messages; fix item loot "x<N>" color
This flips around the icon and value order of money loot text
(1[G] 23[S] 45[C] instead of [G]1 [S]23 [C]45), but the new order is
consistent with Blizzard UI so that's probably for the best.

Blizzard utility scripts also provide CreateTextureMarkup()
(in Interface/SharedXML/TextureUtil.lua) which we could use for
embedding custom icons, but we do still need our own icon element
for buff/debuff masks, which can't be embedded.  The markup could
be useful in damage fly text once we have spell school icons.