Fix attacking enemies sometimes not appearing on enmity list
Caused by expecting UnitIsUnit() to accept a GUID when it does not.
This particularly affects the minor Aqir enemies in the Hive Mind wing
of the Ny'alotha raid, since those generally don't get full nameplates
even when you attack them.

We still can't get HP for enemies without associated tokens, so the
health bar is hidden while an enemy is tokenless.