Add a utility routine to apply the shared UI texture or a part thereof
No named atlas definitions along the lines of SetAtlas() because those
are still overkill at this point, and there doesn't seem to be a way to
define custom atlases for use by SetAtlas() anyway (not that we couldn't
just write our own version).

Also fix legacy backslashes in pathnames to proper slashes.