Possible solution to StaticPopup taint in ItemUpgradeFrame
The game only seems to complain about taint if we fetch the button ID
from the WoWXIV_config table and not if we pass a literal string to
SetOverrideBinding(), so we use low-order bits of the camera speed
cvars to hold the confirm and cancel button IDs.

It's unclear if this is intended taint behavior, but confirm/cancel
ought to have their own cvars anyway, so we'll call it good.