Fix SL mission follower list frame not activating when scrolled
Delayed SetTargets() meant the frame started out with no targets,
so the cursor was not displayed, and a subsequent SetTarget() didn't
scroll (because the cursor was hidden) so the target didn't get a
frame and UpdateCursor() cleared it.

This might fix some other cases of lost cursor on scroll boxes too.
Hopefully it doesn't break anything.