Fill stacks when possible in bank item deposit/withdrawal
If the target bag has a non-full stack of the item being moved but the
entire source stack won't fit, move only enough units to fill the target
stack (like FFXIV) rather than creating a second non-full stack (default
WoW behavior).  For example, given an item with max stack size 100, when
transferring a stack of 70 to a bag with a stack of 60, we now transfer
just 40, giving a full target stack of 100 and leaving 30 in the source
stack, rather than having two stacks of 70 and 60 in the target bag.
A second transfer of the same item would move the remaining 30 into a
new stack in the target (or some other relevant) bag.

It would be even better to send to multiple target stacks so we consume
the entire source stack, but that can't be done instantly and would need
various locking and safety checks, so we pass for now.