Refactor autoloot bug workaround
Reduce LootSlot() calls, fix the workaround for the second form of
the bug, and add a workaround for a _third_ form.  ("This isn't even
my final form!")  Also apply the workaround regardless of whether
loot frame hiding is enabled, and add event triggers so we can more
easily track when the workaround is activated.

There's probably an argument for moving the workaround logic out to
its own module at some point.