Implement __super() for all methods in classes
Not sure why this approach didn't occur sooner; I'll chalk it up to
lack of Lua experience.

Plain "super" (without the underscores, like Python) is another option,
but usage would conflict with Python (which only returns the class
itself) and potentially confuse users.  "__super" is also easier to
unambiguously find-and-replace later on if desired.

Also tweak test runner to run tests in source line order when the
debug library is available.