Remove menu cursor focus while fishing
Allows a confirm button press to hook the fish, assuming the confirm
button is also set as the interact button.

This relies on an aura only available via the combat log (not included
in aura data returned from C_UnitAuras), so the logic may need to be
revisited (or we may have to give up on this) when the combat addon purge
arrives in 12.0.0.